+++ src/TestCentric/testcentric.gui/Presenters/NUnitTreeDisplayStrategy.cs
@@ -47,31 +47,6 @@ namespace TestCentric.Gui.Presenters
             TrySaveVisualState();
         }
 
-        private bool TryRestoreVisualState()
-        {
-            if (_model.TestFiles.Count == 0)
-                return false;
-
-            var filename = VisualState.GetVisualStateFileName(_model.TestFiles[0]);
-            if (!File.Exists(filename))
-                return false;
-
-            var visualState = VisualState.LoadFrom(filename);
-
-            visualState.ApplyTo(_view.TreeView);
-
-            return true;
-        }
-
-        private void TrySaveVisualState()
-        {
-            if (_model.TestFiles.Count > 0)
-            {
-                var visualState = new VisualState().LoadFrom(_view.TreeView);
-                visualState.Save(VisualState.GetVisualStateFileName(_model.TestFiles[0]));
-            }
-        }
-
         private void SetDefaultInitialExpansion()
         {
             var displayStyle = (InitialTreeExpansion)_settings.Gui.TestTree.InitialTreeDisplay;
